connectors/cherrypy-rest/requirements.txt
cherrypy
paho-mqtt

